Fix: open in browser command.
Without this commit, using the `b` command to open Toggl in the browser
would fail by having the browser opening a 404 page.

This commit involves a very minor rearchitecting. Previously, there was
one URL "www.toggl.com" that could be used for both the JSON-based API
endpoint, and the for-humans web browser endpoint. Toggl has since split
this into "www.toggl.com" for humans and "api.track.toggl.com" for API
access.

The concept of the `URL` thus no longer makes sense, and I've removed
that. The concepts of the `API_URL` and the `TIMER_URL` remain.

I've updated the tests to not assert the existence of the `URL`
definition, under the reasoning that this is an implementation detail
that should be free to change, and not part of the `src.toggl.js` file's
public API. Similarly, I've changed the tests on the `API_URL` and
`TIMER_URL` constants to not assert that the URL has a specific value
(e.g. that it contains "/api") because that too is an implementation
detail that may change (if Toggl decides to change its URLs around).
Instead, the tests just assert that we're using an https connection
(although that too may be an implementation detail? No other code relies
on that property, but perhaps the maintainers would like to be alerted
via a test failure if that ever changes...)

Testing done:
* `npm start b`: Confirmed that using the command line to open the
  browser now works.
* `npm start`: Confirmed that pressing `b` in the interactive UI to open
  the browser no works.
* `npm test`: Confirmed all test pass.