chore: Miscelanous fix for xo linter.
This is in preparation to upgrade the `xo` linter from 0.16 to 0.44. The
new version of the linter enforces several new things. This commit
contains a bunch of the "easy fixes" that the new linter enforces, so
that the difficult changes (e.g. switching from Promises to await/async)
can be done in their own commit.

Testing done:
* `npm test` confirmed tests still pass.
* `npm start` confirmed interactive UI loads data.
* `npm start list` confirmed CLI UI loads data.