chore: Upgraded dev dependencies chai 3.5.0 -> 4.3.4, chai-aspromised 5.3.0 -> 7.1.1, chai-spies 0.7.1 -> 1.0.0.
This is in preparation to upgrade the `xo` linter. One of the
enforcements in the newest version of `xo` is that you should use
`async/await` instead of promises. Older versions of `chai` incorrectly
consider async functions to not be a function (i.e. `let f = async
function() {}; f.should.be.a('function');` fails).

Upgrading chai and related libraries allows the test to continue passing
after we change some functions to use the async/await style instead of
directly returning promises.

Testing done:
* `npm test` confirmed all tests still pass.